+++ include/littlehell/net/client.h
@@ -20,7 +20,7 @@
 #include "doomtype.h"
 #include "d_ticcmd.h"
 #include "sha1.h"
-#include "net_defs.h"
+#include "defs.h"
 
 boolean NET_CL_Connect(net_addr_t *addr, net_connect_data_t *data);
 void NET_CL_Disconnect(void);
