+++ src/net/loop.c
