+++ src/net/query.c
