+++ src/net/structrw.c
