+++ src/net/structrw.h
